tabella quasi ok...

+++ net_builder_widget.cpp
@@ -1,7 +1,6 @@
 #include "net_builder_widget.h"
 #include <QLayoutItem>
 #include <QLabel>
-#include <QPalette>
 
 netBuilderWidget::netBuilderWidget(QWidget * parent) : QWidget(parent), layout(new QFormLayout), name(new QLineEdit), neuron_type(new QComboBox), n_layers(new QSpinBox), next(new QPushButton("Prosegui...")), trigger(new QPushButton("Crea")), new_net(nullptr)
 {
@@ -38,14 +37,12 @@ void netBuilderWidget::setLayers()
 		new_net = new layeredTanhNet;
 	else if (type == "Arcotangente")
 		new_net = new layeredAtanNet;
-//	else ... default ... eccezione
-
-//	emit newNet(name->text(), neuron_type->currentText(), n_layers->value());
+//	else... nullptr
 
 	QSpacerItem * spacer = new QSpacerItem(1, 20);
 	layout->addItem(spacer);
 	layout->addRow("Numero di unità per livello:", new QLabel);
-	for (unsigned int i = 0; i < static_cast<unsigned int>(n_layers->value()); ++i)
+	for (int i = 0; i < n_layers->value(); ++i)
 	{
 		layers_arch.push_back(new QSpinBox);
 		layers_arch[i]->setMinimum(1);
@@ -64,7 +61,7 @@ void netBuilderWidget::addNet()
 
 	if (new_net)
 	{
-		for (unsigned int i = 0; i < static_cast<unsigned int>(n_layers->value()); ++i)
+		for (int i = 0; i < n_layers->value(); ++i)
 			new_net->addLayer(layers_arch[i]->value());
 
 		new_net->init(0.5f);
